coffee_project/requirements.txt
googlemaps>=4.10.0
pandas>=1.5.0
geopandas>=0.14.0
python-dotenv>=1.0.0
requests>=2.28.0
anthropic>=0.21.0 